Use bfloat16 where possible

+++ examples/basic_usage.py
@@ -22,8 +22,8 @@ print(f"  hidden_size = {hidden_size:,}")
 print(f"  n_hidden_states = {n_hidden_states:,}")
 
 # Create random inputs
-weights = torch.randn(vocab_size, hidden_size, dtype=torch.float32)
-hidden_states = torch.randn(hidden_size, n_hidden_states, dtype=torch.float32)
+weights = torch.randn(vocab_size, hidden_size, dtype=torch.bfloat16)
+hidden_states = torch.randn(hidden_size, n_hidden_states, dtype=torch.bfloat16)
 
 # Sample from categorical distribution using fused Triton kernel
 samples = fused_mm_sample_triton(
